requirements.txt
boto==2.34.0
django-celery-with-redis==3.0
django-allauth==0.18.0
django-analytical==0.19.0
django-annoying==0.8.0
django-compressor==1.4
django-heroku-memcacheify==0.7
django-heroku-postgresify==0.3
django-extensions==1.4.4
django-nose==1.2
django-selector==0.4
django-staticfiles==1.2.1
django-storages==1.1.8
django-ses==0.6.0
gunicorn==19.1.1
requests==2.4.3
psycopg2==2.5.4
sorl-thumbnail==11.12
splinter==0.6.0
versiontools==1.9.1

# Dev only
django-debug-toolbar==1.2.1
ipython==2.3.0
## The following requirements were added by pip --freeze:
BeautifulSoup==3.2.1
Django==1.5.11
Fabric==1.10.0
Pillow==2.6.1
South==1.0.1
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.16
dj-database-url==0.1.2
django-appconf==0.6
django-celery==3.1.16
django-pylibmc==0.5.0
ecdsa==0.11
gnureadline==6.3.3
kombu==3.0.23
nose==1.3.4
oauthlib==0.7.2
paramiko==1.15.1
pycrypto==2.6.1
pylibmc==1.3.0
python-openid==2.2.5
pytz==2014.7
redis==2.10.3
requests-oauthlib==0.4.2
selenium==2.44.0
six==1.8.0
sqlparse==0.1.13
wsgiref==0.1.2
